Everything working (che sbatti) till beginning of 131
