+++ backend/app.js
@@ -19,6 +19,7 @@ mongoose.connect("mongodb+srv://mongo-adm:qDYNoySK2GiZ44tZ@cluster0-lb7gt.mongod
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
+
 app.use("/images", express.static(path.join("backend/images")));
 app.use(cors());
 
@@ -26,7 +27,7 @@ app.use((_req, res, next) => {
   res.setHeader(
     "Access-Control-Allow-Origin", "*");
 
- res.setHeader(
+  res.setHeader(
     "Access-Control-Allow-Headers",
     "Origin, X-Requested-With, Content-Type, Accept, Authorization,"
     );
