+++ package.json
@@ -32,6 +32,7 @@
     "mongoose": "^5.5.9",
     "mongoose-unique-validator": "^2.0.3",
     "multer": "^1.4.2",
+    "req": "^0.1.4",
     "rxjs": "~6.3.3",
     "tslib": "^1.10.0",
     "zone.js": "~0.8.26"
